Moved some more files to the correct place.
