+++ app/Http/Controllers/Cooperation/Frontend/Tool/SimpleScan/QuestionnaireController.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace App\Http\Controllers\Cooperation\Frontend\Tool\QuickScan;
+namespace App\Http\Controllers\Cooperation\Frontend\Tool\SimpleScan;
 
+use App\Http\Controllers\Controller;
 use App\Models\Cooperation;
 use App\Models\Questionnaire;
 use App\Models\Scan;
 use App\Models\Step;
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 
 class QuestionnaireController extends Controller
 {
@@ -16,6 +15,6 @@ class QuestionnaireController extends Controller
         // Ensure it's a valid questionnaire
         abort_if($questionnaire->isNotActive() || $questionnaire->step->id !== $step->id, 404);
 
-        return view('cooperation.frontend.tool.quick-scan.questionnaires.index', compact('step', 'questionnaire'));
+        return view('cooperation.frontend.tool.simple-scan.questionnaires.index', compact('step', 'questionnaire'));
     }
 }
